Module rocket_okapi::handlers

Contains several Rocket Handlers, which are used for serving the json files and the swagger interface.

Structs

ContentHandler
A content handler is a wrapper type around rocket::response::Content, which can be turned into a rocket::Route that serves the content with correct content-type.

OpenApiHandler
A handler type that is used to serve the openapi.json files.

RedirectHandler
A handler that instead of serving content always redirects to some specified destination URL.