Trait rocket_okapi::request::OpenApiFromParam

pub trait OpenApiFromParam<'r>: FromParam<'r> {
    fn path_parameter(
        gen: &mut OpenApiGenerator, 
        name: String
    ) -> Result<Parameter>;
}

This trait means that the implementer can be used as a FromParam request guard, and that this can also be documented.

Required methods

fn path_parameter(gen: &mut OpenApiGenerator, name: String) -> Result<Parameter>

Return a RequestBody containing the information required to document the FromParam for implementer.

Loading content...