Trait rocket_okapi::request::OpenApiFromQuery

pub trait OpenApiFromQuery<'r>: FromQuery<'r> {
    fn query_multi_parameter(
        gen: &mut OpenApiGenerator, 
        name: String
    ) -> Result<Parameter>;
}

This trait means that the implementer can be used as a FromQuery request guard, and that this can also be documented.

Required methods

fn query_multi_parameter(
gen: &mut OpenApiGenerator,
name: String
) -> Result<Parameter>

Return a RequestBody containing the information required to document the FromQuery for implementer.

Loading content...