Trait rocket_okapi::response::OpenApiResponder

pub trait OpenApiResponder<'r>: Responder<'r> {
    fn responses(gen: &mut OpenApiGenerator) -> Result<Responses>;
}

Implementing this trait means that any route returning the implementer can be marked with #[openapi], and that the route can be documented.

Required methods

fn responses(gen: &mut OpenApiGenerator) -> Result<Responses>

Create the responses type, which is a list of responses that can be rendered in openapi.json format.

Loading content...